Fixed issue with theme selection
Isse exitsted because the theme session var is set in / so when you
clicked on HOME link it checked to see if the post was set, since it was
not it set the ThemeSelected var to false meaning it reset to default.
added an empty if statment that jumps over the else clause and now it
doesnt reset until you tell it too.